feat: [PIE-4290]: Template gitx integration (#10274)
* feat: [PIE-4290]: Template listing page and creation modal, save modal WIP

* feat: [PIE-4290]: Fix template cache idb issue

* feat: [PIE-4290]: Save as template from pipeline studio new git exp

* feat: [PIE-4290]: Resolve issues after rebase

* feat: [PIE-4290]: Save as new version WIP

* feat: [PIE-4290]: Add version label to filePath

* feat: [PIE-4290]: WIP

* feat: [PIE-4290]: Show error on template studio

* feat: [PIE-4290]: Use existing template error

* feat: [PIE-4290]: Minor changes

* feat: [PIE-4290]: Remove git fields from initial values

* feat: [PIE-4290]: Remove showGitFields

* feat: [PIE-4290]: Use intent to determine behavior

* feat: [PIE-4290]: Use intent to determine comment behavior

* feat: [PIE-4290]: Fix inline template api request

* feat: [PIE-4290]: Fix inline pipeline to save as remote template

* feat: [PIE-4290]: Cleanup template subheader onsubmit

* feat: [PIE-4290]: Write a separate function for inline remote select

* feat: [PIE-4290]: Expose only fetchTemplate from TemplateContext

* feat: [PIE-4290]: Update deps array in hook

* feat: [PIE-4290]: Fix inline template save as new remote

* feat: [PIE-4290]: Fix undefined error

* feat: [PIE-4290]: Add entityScope to GitSyncForm

* feat: [PIE-4290]: Remove TemplateError component

* feat: [PIE-4290]: Use list metadata api for template settings

* feat: [PIE-4290]: Fix git details invisible during template creation

* feat: [PIE-4290]: Template selector changes

* feat: [PIE-4290]: Pass parent entity context to template selector

* feat: [PIE-4290]: Use new preject settings variable for git simplification

* feat: [PIE-4290]: Fix issues after pull

* feat: [PIE-4290]: Fix yaml tab error exp

* feat: [PIE-4290]: Use new FF for templates git exp

* feat: [PIE-4290]: Fix some failing tests

* feat: [PIE-4290]: Fix some sonar issues

* feat: [PIE-4290]: Fix type error

* feat: [PIE-4290]: Fix ternary check sonar

* feat: [PIE-4290]: Fix call retries exceeded error

* feat: [PIE-4290]: Fix cypress failing test

* feat: [PIE-4290]: Add test case for useParentEntityContext hook

* feat: [PIE-4290]: Fix create PR issue

* feat: [PIE-4290]: Pass parent entity branch in GET call

* feat: [PIE-4290]: Resolve review comments

* feat: [PIE-4290]: Fix template config modal scroll issue

* feat: [PIE-4290]: Remove isInitialized from template context action

* feat: [PIE-4290]: Add FF to disabled remote template creation at account/org scope

* feat: [PIE-4290]: Minor refactor

* feat: [PIE-4290]: Use template account ff to disable remote on scope change

* feat: [PIE-4290]: Refactor useSaveAsTemplate

* feat: [PIE-4290]: Init db

* feat: [PIE-4290]: prevent get yaml call when template gitx disabled

* feat: [PIE-4290]: Pass storeMetadata to PipelineStages

* feat: [PIE-4290]: Pass storeMetadata in selectorData for template selector

* feat: [PIE-4290]: Delete obsolete files

* feat: [PIE-4290]: Refactor TemplateDetails

* feat: [PIE-4290]: Allow branch change from template preview

* feat: [PIE-4290]: Fix minor issues

* feat: [PIE-4290]: Simplify condition

* feat: [PIE-4290]: Fix template api issue

* feat: [PIE-4290]: Fix git popover branch issue

* feat: [PIE-4290]: Refactor template config modal

* feat: [PIE-4290]: Revert a line

* feat: [PIE-4290]: Refactor WIP

* feat: [PIE-4290]: Refactor PipelineStages

* feat: [PIE-4290]: Revert some code

* feat: [PIE-4290]: Sonar fixes

* feat: [PIE-4290]: Fix test cases

* feat: [PIE-4290]: Fix failing cypress test case

+++ src/modules/72-templates-library/components/TemplateStudio/TemplateStudio.tsx
@@ -77,9 +77,12 @@ export function TemplateStudio(): React.ReactElement {
     isBETemplateUpdated,
     isInitialized,
     gitDetails,
+    storeMetadata,
     entityValidityDetails,
-    templateInputsErrorNodeSummary
+    templateInputsErrorNodeSummary,
+    templateYamlError
   } = state
+
   const { isYamlEditable } = templateView
   const { getString } = useStrings()
   const [blockNavigation, setBlockNavigation] = React.useState(false)
@@ -187,42 +190,45 @@ export function TemplateStudio(): React.ReactElement {
   }, [isBETemplateUpdated, discardBEUpdateDialog, openConfirmBEUpdateError, blockNavigation, openUnsavedChangesDialog])
 
   const onGitBranchChange = React.useMemo(
-    () => (selectedFilter: GitFilterScope) => {
+    () => (selectedFilter: GitFilterScope, defaultSelected?: boolean) => {
       setSelectedBranch(selectedFilter.branch as string)
-      if (isUpdated && branch !== selectedFilter.branch) {
-        setBlockNavigation(true)
-      } else if (branch !== selectedFilter.branch) {
-        deleteTemplateCache({
-          repoIdentifier: defaultTo(selectedFilter.repo, ''),
-          branch: defaultTo(selectedFilter.branch, '')
-        }).then(() => {
-          history.push(
-            routes.toTemplateStudio({
-              projectIdentifier,
-              orgIdentifier,
-              templateIdentifier: defaultTo(templateIdentifier, '-1'),
-              accountId,
-              module,
-              templateType: template.type,
-              versionLabel: template.versionLabel,
-              branch: selectedFilter.branch,
-              repoIdentifier: selectedFilter.repo
-            })
-          )
-        })
+      if (!defaultSelected) {
+        if (isUpdated && branch !== selectedFilter.branch) {
+          setBlockNavigation(true)
+        } else if (branch !== selectedFilter.branch) {
+          deleteTemplateCache({
+            repoIdentifier: defaultTo(selectedFilter.repo, ''),
+            branch: defaultTo(selectedFilter.branch, '')
+          }).then(() => {
+            history.push(
+              routes.toTemplateStudio({
+                projectIdentifier,
+                orgIdentifier,
+                templateIdentifier: defaultTo(templateIdentifier, '-1'),
+                accountId,
+                module,
+                templateType: template.type,
+                versionLabel: template.versionLabel,
+                branch: selectedFilter.branch,
+                repoIdentifier: selectedFilter.repo
+              })
+            )
+          })
+        }
       }
     },
     [
-      branch,
-      isUpdated,
-      templateIdentifier,
-      projectIdentifier,
-      orgIdentifier,
       accountId,
-      module,
+      branch,
       deleteTemplateCache,
       history,
-      template
+      isUpdated,
+      module,
+      orgIdentifier,
+      projectIdentifier,
+      template.type,
+      template.versionLabel,
+      templateIdentifier
     ]
   )
 
@@ -253,6 +259,50 @@ export function TemplateStudio(): React.ReactElement {
       await templateStudioSubHeaderHandleRef.current?.updateTemplate(entityYaml)
     },
     [templateStudioSubHeaderHandleRef.current]
+  )
+
+  const ErrorPanel = (): JSX.Element => (
+    <Container style={{ maxWidth: '570px', alignSelf: 'center' }} padding={'huge'}>
+      <NoEntityFound
+        identifier={templateIdentifier}
+        entityType={'template'}
+        errorObj={templateYamlError}
+        gitDetails={{
+          connectorRef: storeMetadata?.connectorRef,
+          repoName: storeMetadata?.repoName,
+          branch: storeMetadata?.branch,
+          onBranchChange: onGitBranchChange
+        }}
+      />
+    </Container>
+  )
+
+  const renderView =
+    view === SelectedView.VISUAL ? (
+      /* istanbul ignore next */
+      templateFactory.getTemplate(templateType)?.renderTemplateCanvas(templateFormikRef)
+    ) : (
+      <TemplateYamlView />
+    )
+
+  const studioCanvasUI = templateYamlError ? (
+    <ErrorPanel />
+  ) : (
+    <>
+      {templateInputsErrorNodeSummary && (
+        <OutOfSyncErrorStrip
+          errorNodeSummary={templateInputsErrorNodeSummary}
+          entity={TemplateErrorEntity.TEMPLATE}
+          originalYaml={yamlStringify({ template: originalTemplate })}
+          isReadOnly={isReadonly}
+          onRefreshEntity={() => {
+            fetchTemplate({ forceFetch: true, forceUpdate: true })
+          }}
+          updateRootEntity={updateEntity}
+        />
+      )}
+      <Container className={css.canvasContainer}>{renderView}</Container>
+    </>
   )
 
   return (
@@ -306,25 +356,7 @@ export function TemplateStudio(): React.ReactElement {
                   getErrors={getErrors}
                   onGitBranchChange={onGitBranchChange}
                 />
-                {templateInputsErrorNodeSummary && (
-                  <OutOfSyncErrorStrip
-                    errorNodeSummary={templateInputsErrorNodeSummary}
-                    entity={TemplateErrorEntity.TEMPLATE}
-                    originalYaml={yamlStringify({ template: originalTemplate })}
-                    isReadOnly={isReadonly}
-                    onRefreshEntity={() => {
-                      fetchTemplate({ forceFetch: true, forceUpdate: true })
-                    }}
-                    updateRootEntity={updateEntity}
-                  />
-                )}
-                <Container className={css.canvasContainer}>
-                  {view === SelectedView.VISUAL ? (
-                    templateFactory.getTemplate(templateType)?.renderTemplateCanvas(templateFormikRef)
-                  ) : (
-                    <TemplateYamlView />
-                  )}
-                </Container>
+                {studioCanvasUI}
               </>
             )}
             {templateType !== TemplateType.Pipeline && <RightBar />}
